fix memory leak in VulkanSamples/loader/loader.c loader_destroy_layer_list()
This one is easy to reproduce:
Run valgrind --leak_check=full on a very short app that does:
int main() {
  uint32_t layerCount = 0;
  vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
  printf(layerCount=%un, layerCount);
  return 0;
}

valgrind should report:
==PID==
==PID== 49,920 bytes in 6 blocks are definitely lost in loss record 57 of 58
==PID==    at 0x40307C4: malloc (vg_replace_malloc.c:270)
==PID==    by 0x4A2B5F3: loader_instance_heap_alloc (VulkanSamples/loader/loader.c:178)
==PID==    by 0x4A2CB0D: loader_init_generic_list (VulkanSamples/loader/loader.c:858)
==PID==    by 0x4A2CBDE: loader_add_to_ext_list (VulkanSamples/loader/loader.c:889)
==PID==    by 0x4A30A32: loader_read_json_layer (VulkanSamples/loader/loader.c:2311)
==PID==    by 0x4A31373: loader_add_layer_properties (VulkanSamples/loader/loader.c:2509)
==PID==    by 0x4A3286F: loader_layer_scan (VulkanSamples/loader/loader.c:3032)
==PID==    by 0x4A38DFC: vkEnumerateInstanceLayerProperties (VulkanSamples/loader/trampoline.c:239)
==PID==

The problem is that loader_destroy_layer_list() does not do the cleanup steps found in loader_delete_layer_properties().
It appears the two are really the same function implemented twice and a small refactor is needed.
But a quick fix is to just copy-paste the parts of loader_delete_layer_properties() that loader_destroy_layer_list() needs.